Test code for passing tools to python functions
I have added some testing code demonstrating how to pass CP tools to
user-specified python functions for configuration. The error checking
within the python code is still crude, but this could be cleaned up a
bit by pythonizing the tools and catching errors in the property getters
and setters.

Requires ROOT dictionaries for the tools.